fix shell commands
